requirements.txt
flask
numpy
openai
pandas
python-dotenv
tiktoken